changed angular version to 19.2.0
