+++ task-manager/tsconfig.app.json
@@ -1,4 +1,5 @@
-/* To learn more about this file see: https://angular.io/config/tsconfig. */
+/* To learn more about Typescript configuration file: https://www.typescriptlang.org/docs/handbook/tsconfig-json.html. */
+/* To learn more about Angular compiler options: https://angular.dev/reference/configs/angular-compiler-options. */
 {
   "extends": "./tsconfig.json",
   "compilerOptions": {
@@ -10,7 +11,7 @@
   "files": [
     "src/main.ts",
     "src/main.server.ts",
-    "server.ts"
+    "src/server.ts"
   ],
   "include": [
     "src/**/*.d.ts"
